+++ lib/fieldhand/list_records_parser.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'fieldhand/record'
 
 module Fieldhand