+++ lib/fieldhand/repository.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'fieldhand/arguments'
 require 'fieldhand/get_record_parser'
 require 'fieldhand/identify_parser'
@@ -7,6 +9,7 @@ require 'fieldhand/list_records_parser'
 require 'fieldhand/list_sets_parser'
 require 'fieldhand/options'
 require 'fieldhand/paginator'
+require 'forwardable'
 require 'uri'
 
 module Fieldhand
@@ -14,23 +17,26 @@ module Fieldhand
   #
   # See https://www.openarchives.org/OAI/openarchivesprotocol.html
   class Repository
-    attr_reader :uri, :logger, :timeout, :headers
+    attr_reader :uri, :logger_or_options
 
-    # Return a new repository with the given base URL and an optional logger, timeout, bearer token and headers.
+    # Allow the user to query the repository's current options.
+    extend Forwardable
+    def_delegators :paginator, :logger, :timeout, :retries, :interval, :headers
+
+    # Return a new repository with the given base URL and an optional logger,
+    # timeout, maximum number of retries, retry interval, bearer token and
+    # headers.
     #
     # The base URL can be passed as a `URI` or anything that can be parsed as a URI such as a string.
     #
     # For backward compatibility, the second argument can either be a logger or a hash containing
-    # a logger, timeout, bearer token and headers.
+    # a logger, timeout, maximum number of retries, retry interval, bearer token and headers.
     #
-    # Defaults to using a null logger specific to this platform, a timeout of 60 seconds, no bearer token and no headers.
+    # Defaults to using a null logger specific to this platform, a timeout of 60 seconds, a maximum
+    # number of retries of 0, a retry interval of 10 seconds, no bearer token and no headers.
     def initialize(uri, logger_or_options = {})
       @uri = uri.is_a?(::URI) ? uri : URI(uri)
-
-      options = Options.new(logger_or_options)
-      @logger = options.logger
-      @timeout = options.timeout
-      @headers = options.headers
+      @logger_or_options = logger_or_options
     end
 
     # Send an Identify request to the repository and return an `Identify` response.
@@ -134,7 +140,7 @@ module Fieldhand
     private
 
     def paginator
-      @paginator ||= Paginator.new(uri, :logger => logger, :timeout => timeout, :headers => headers)
+      @paginator ||= Paginator.new(uri, logger_or_options)
     end
   end
 end